refactor(attention): consolidate KPool indexer backend
Keep KPool cache metadata and model execution in one attention module. Update model and test imports and defer KPool operator loading until backend initialization.
